Fixed the bugs.
