+++ HAL/Android.bp
@@ -22,6 +22,9 @@ cc_library {
     	"CborConverter.cpp",
         "JavacardKeyMintDevice.cpp",
         "JavacardKeyMintOperation.cpp",
+        "JavacardSecureElement.cpp",
+        "JavacardSharedSecret.cpp",
+        "keymint_utils.cpp",
     ],
     cflags:["-O0",],
     shared_libs: [
@@ -90,7 +93,6 @@ cc_binary {
     ],
     srcs: [
         "service.cpp",
-        "keymint_utils.cpp"
     ],
     required: [
 //        "RemoteProvisioner",
