+++ HAL/JavacardSharedSecret.h
@@ -1,13 +1,15 @@
 #pragma once
 
-#include "JavacardKeyMintDevice.h"
+#include "CborConverter.h"
+#include "JavacardSecureElement.h"
+
 #include <aidl/android/hardware/security/sharedsecret/BnSharedSecret.h>
 #include <aidl/android/hardware/security/sharedsecret/SharedSecretParameters.h>
 #include <memory>
 #include <vector>
 
 namespace aidl::android::hardware::security::sharedsecret {
-using namespace aidl::android::hardware::security::keymint;
+using namespace ::keymint::javacard;
 using ndk::ScopedAStatus;
 using std::optional;
 using std::shared_ptr;
@@ -15,20 +17,17 @@ using std::vector;
 
 class JavacardSharedSecret : public BnSharedSecret {
   public:
-    explicit JavacardSharedSecret(shared_ptr<JavacardKeyMintDevice> keymint) : keymint_(keymint) {}
+    explicit JavacardSharedSecret(shared_ptr<JavacardSecureElement> card) : card_(card) {}
     virtual ~JavacardSharedSecret() {}
 
-    ScopedAStatus getSharedSecretParameters(SharedSecretParameters* params) override {
-        return keymint_->getSharedSecretParameters(params);
-    }
+    ScopedAStatus getSharedSecretParameters(SharedSecretParameters* params) override;
 
     ScopedAStatus computeSharedSecret(const std::vector<SharedSecretParameters>& params,
-                                      std::vector<uint8_t>* secret) override {
-        return keymint_->computeSharedSecret(params, secret);
-    }
+                                      std::vector<uint8_t>* secret) override;
 
   private:
-    shared_ptr<JavacardKeyMintDevice> keymint_;
+    shared_ptr<JavacardSecureElement> card_;
+    CborConverter cbor_;
 };
 
 }  // namespace aidl::android::hardware::security::sharedsecret
