+++ HAL/service.cpp
@@ -23,8 +23,9 @@
 #include "JavacardKeyMintDevice.h"
 #include <aidl/android/hardware/security/keymint/SecurityLevel.h>
 
-#include "keymint_utils.h"
+#include "JavacardSecureElement.h"
 #include "JavacardSharedSecret.h"
+#include "keymint_utils.h"
 //#include "JavacardRemotelyProvisionedComponentDevice.h"
 #include <SocketTransport.h>
 
@@ -45,16 +46,17 @@ template <typename T, class... Args> std::shared_ptr<T> addService(Args&&... arg
 
 int main() {
     ABinderProcess_setThreadPoolMaxThreadCount(0);
+    // Javacard Secure Element
+    std::shared_ptr<JavacardSecureElement> card =
+        std::make_shared<JavacardSecureElement>(std::make_shared<SocketTransport>(), getOsVersion(),
+                                                getOsPatchlevel(), getVendorPatchlevel());
     // Add Keymint Service
-
-    std::shared_ptr<JavacardKeyMintDevice> keyMint =
-        addService<JavacardKeyMintDevice>(std::make_shared<SocketTransport>(), getOsVersion(),
-                                          getOsPatchlevel(), getVendorPatchlevel());
+    addService<JavacardKeyMintDevice>(card);
     // Add Shared Secret Service
-        addService<JavacardSharedSecret>(keyMint);
+    addService<JavacardSharedSecret>(card);
     // Add Remotely Provisioned Component Service
-      //  addService<JavacardRemotelyProvisionedComponentDevice>(keyMint);
-        
+    //  addService<JavacardRemotelyProvisionedComponentDevice>(keyMint);
+
     ABinderProcess_joinThreadPool();
     return EXIT_FAILURE;  // should not reach
 }
